Move TLSConfig from provider to types in the way to avoid cyclic dependencies.
