+++ provider/kv/kv.go
@@ -23,7 +23,7 @@ type Provider struct {
 	provider.BaseProvider `mapstructure:",squash"`
 	Endpoint              string              `description:"Comma separated server endpoints"`
 	Prefix                string              `description:"Prefix used for KV store"`
-	TLS                   *provider.ClientTLS `description:"Enable TLS support"`
+	TLS                   *types.ClientTLS `description:"Enable TLS support"`
 	Username              string              `description:"KV Username"`
 	Password              string              `description:"KV Password"`
 	storeType             store.Backend
